Still getting undefined index errors
I completely missed the fact that, when I fixed the first undefined index error that actually more would follow.

$_GET['steamids'] is used three times in the last block of code.

I have wrapped that all in a conditional so that doesn't execute at all. Also before that I've set the $content_json variable as an empty string, just in case there are no steamIds, it will just echo blank but please check if that is desired behavior.